Move _create_message_box and _create_confirmation_box to /gui/components.py
